requirements.txt
flask
pymupdf
lxml
pillow

